+++ Changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v4.5.0
+
+- Handle multiple cash accounts.
+- Able to see past year dividend bar chart.
+
 ## v4.4.1
 
 - Added account filters and price column to overview tables.